third_party/blink/web_tests/fast/events/touch/gesture/gesture-dblclick-expected.txt

This tests gesture callbacks for a double tap sequence of gestures.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS event.type is "mousemove"
PASS event.clientX is 10
PASS event.clientY is 12
PASS event.detail is 0
PASS event.type is "mousedown"
PASS event.clientX is 10
PASS event.clientY is 12
PASS event.detail is 1
PASS event.type is "mouseup"
PASS event.clientX is 10
PASS event.clientY is 12
PASS event.detail is 1
PASS event.type is "click"
PASS event.clientX is 10
PASS event.clientY is 12
PASS event.detail is 1
PASS event.type is "mousemove"
PASS event.clientX is 11
PASS event.clientY is 12
PASS event.detail is 0
PASS event.type is "mousedown"
PASS event.clientX is 11
PASS event.clientY is 12
PASS event.detail is 2
PASS event.type is "mouseup"
PASS event.clientX is 11
PASS event.clientY is 12
PASS event.detail is 2
PASS event.type is "click"
PASS event.clientX is 11
PASS event.clientY is 12
PASS event.detail is 2
PASS event.type is "dblclick"
PASS event.clientX is 11
PASS event.clientY is 12
PASS event.detail is 2
PASS successfullyParsed is true

TEST COMPLETE
